update tds to float

+++ WaterQuality.Domain/Water/WaterQualityParameter.cs
@@ -1,2 +1,2 @@
 ﻿namespace WaterQuality.Domain.Water;
-public record WaterQualityParameter(int Tds = 0, decimal Temperature = 0, decimal Ph = 0, DateTime Date = default);
+public record WaterQualityParameter(float Tds = 0, decimal Temperature = 0, decimal Ph = 0, DateTime Date = default);
